Degree Distributions-Clustering Coefficient
Download the DBLP co-authorship matrix and the vertex labels from Moodle. This dataset is a co-authorship graph of computer scientists, where each vertex corresponds to an author and there is an edge if two authors have co-authored a paper. File dblp-co-authors.txt provides the adjacency matrix in so-called coordinate format: Each row corresponds to one non-zero entry in the matrix. The rows have structure “i j v”, where i is the row number (starting from 1), j is the column number, and v is the value so that A(i, j) = v. For this matrix, v is always 1 and the indices are sorted primarily by row and secondarily by column in ascending order. It was claimed in the lectures that the degree distribution of social networks will (often) follow the power law. Compute the degree distribution of the DBLP data and compute the power law exponent γ that fits best. You can ignore some early part of the data if you feel that it is appropriate, but you must motivate your decision. Plot the degree distribution and your power law curve. Do you think the distribution follows power law? (You must fit the distribution even if you think it does not fit.
Problem 3 (Local Clustering Coefficient). Using the same DBLP graph as in Question 2, compute the average local clustering coefficient for it. Is the network clique-like? How many vertices have local clustering coefficient 1.0? How many have it 0.0